pyproject.toml
[project]
name = "eigen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "ashtyn3", email = "ashtynmorel@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "metalcompute>=0.2.9",
    "metalgpu>=1.0.5",
    "numpy>=2.2.5",
    "pyobjc>=11.0",
    "pyobjc-framework-metal>=11.0",
]

[tool.hatch.build.targets.wheel]
packages = ["eigen/"]

# [tool.hatch.build.targets.wheel]
# include = [
#   "aigan",
# ]

[tool.hatch.build.targets.wheel.data]
"aigan/py.typed" = "eigen/py.typed"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
